TRUNK: Marker API stores only raw counts as uint64_t, evaluation done later. Additionally markerClose supports multipe eventSets and switchMarkerGroup to change eventSet during runtime

+++ src/includes/libperfctr_types.h
@@ -38,15 +38,15 @@ typedef struct LikwidThreadResults{
     double time;
     TimerData startTime;
     uint32_t count;
-    double StartPMcounters[NUM_PMC];
-    double PMcounters[NUM_PMC];
+    uint64_t StartPMcounters[NUM_PMC];
+    uint64_t PMcounters[NUM_PMC];
 } LikwidThreadResults;
 
 typedef struct {
     bstring  tag;
     double*  time;
     uint32_t*  count;
-    double** counters;
+    uint64_t** counters;
 } LikwidResults;
 
 #endif /*LIBPERFCTR_H*/
